Move Arbitrary instances and tests into their own subprojects.
