+++ core/build.sbt
@@ -3,9 +3,3 @@ fork in run := true
 connectInput in run := true
 
 outputStrategy := Some(StdoutOutput)
-
-import scoverage._
-
-ScoverageKeys.coverageMinimum := 57
-
-ScoverageKeys.coverageFailOnMinimum := true
