+++ tests/src/test/scala/matryoshka/spec.scala
@@ -17,6 +17,9 @@
 package matryoshka
 
 import Recursive.ops._, FunctorT.ops._, Fix._
+import matryoshka.exp._
+import matryoshka.exp2._
+import matryoshka.scalacheck.arbitrary._
 
 import java.lang.String
 import scala.{Boolean, Function, Int, None, Option, Predef, Symbol, Unit},
@@ -28,253 +31,31 @@ import org.specs2.ScalaCheck
 import org.specs2.mutable._
 import org.specs2.scalaz._
 import scalaz._, Scalaz._
-import scalaz.scalacheck.ScalaCheckBinding._
-import scalaz.scalacheck.ScalazProperties._
 
-sealed trait Exp[A]
-object Exp {
-  case class Num[A](value: Int) extends Exp[A]
-  case class Mul[A](left: A, right: A) extends Exp[A]
-  case class Var[A](value: Symbol) extends Exp[A]
-  case class Lambda[A](param: Symbol, body: A) extends Exp[A]
-  case class Apply[A](func: A, arg: A) extends Exp[A]
-  case class Let[A](name: Symbol, value: A, inBody: A) extends Exp[A]
-
-  implicit val arbSymbol = Arbitrary(Arbitrary.arbitrary[String].map(Symbol(_)))
-
-  implicit val arbExp: Arbitrary ~> λ[α => Arbitrary[Exp[α]]] =
-    new (Arbitrary ~> λ[α => Arbitrary[Exp[α]]]) {
-      def apply[α](arb: Arbitrary[α]): Arbitrary[Exp[α]] =
-        Arbitrary(Gen.oneOf(
-          Arbitrary.arbitrary[Int].map(Num[α](_)),
-          (arb.arbitrary ⊛ arb.arbitrary)(Mul(_, _)),
-          Arbitrary.arbitrary[Symbol].map(Var[α](_)),
-          (Arbitrary.arbitrary[Symbol] ⊛ arb.arbitrary)(Lambda(_, _)),
-          (arb.arbitrary ⊛ arb.arbitrary)(Apply(_, _)),
-          (Arbitrary.arbitrary[Symbol] ⊛ arb.arbitrary ⊛ arb.arbitrary)(
-            Let(_, _, _))))
-    }
-
-  def num(v: Int) = Fix[Exp](Num(v))
-  def mul(left: Fix[Exp], right: Fix[Exp]) = Fix[Exp](Mul(left, right))
-  def vari(v: Symbol) = Fix[Exp](Var(v))
-  def lam(param: Symbol, body: Fix[Exp]) = Fix[Exp](Lambda(param, body))
-  def ap(func: Fix[Exp], arg: Fix[Exp]) = Fix[Exp](Apply(func, arg))
-  def let(name: Symbol, v: Fix[Exp], inBody: Fix[Exp]) = Fix[Exp](Let(name, v, inBody))
-
-  implicit val ExpTraverse: Traverse[Exp] = new Traverse[Exp] {
-    def traverseImpl[G[_], A, B](fa: Exp[A])(f: A => G[B])(implicit G: Applicative[G]): G[Exp[B]] = fa match {
-      case Num(v)           => G.point(Num(v))
-      case Mul(left, right) => G.apply2(f(left), f(right))(Mul(_, _))
-      case Var(v)           => G.point(Var(v))
-      case Lambda(p, b)     => G.map(f(b))(Lambda(p, _))
-      case Apply(func, arg) => G.apply2(f(func), f(arg))(Apply(_, _))
-      case Let(n, v, i)     => G.apply2(f(v), f(i))(Let(n, _, _))
-    }
-  }
-
-  // NB: an unusual definition of equality, in that only the first 3 characters
-  //     of variable names are significant. This is to distinguish it from `==`
-  //     as well as from a derivable Equal.
-  implicit val ExpEqual: Equal ~> λ[α => Equal[Exp[α]]] =
-    new (Equal ~> λ[α => Equal[Exp[α]]]) {
-      def apply[α](eq: Equal[α]) =
-        Equal.equal[Exp[α]] {
-          case (Num(v1), Num(v2))                 => v1 ≟ v2
-          case (Mul(a1, b1), Mul(a2, b2))         =>
-            eq.equal(a1, a2) && eq.equal(b1, b2)
-          case (Var(s1), Var(s2))                 =>
-            s1.name.substring(0, 3 min s1.name.length) ==
-              s2.name.substring(0, 3 min s2.name.length)
-          case (Lambda(p1, a1), Lambda(p2, a2))   =>
-            p1 == p2 && eq.equal(a1, a2)
-          case (Apply(f1, a1), Apply(f2, a2))     =>
-            eq.equal(f1, f2) && eq.equal(a1, a2)
-          case (Let(n1, v1, i1), Let(n2, v2, i2)) =>
-            n1 == n2 && eq.equal(v1, v2) && eq.equal(i1, i2)
-          case _                                  => false
-        }
-    }
-  implicit def ExpEqual2[A](implicit A: Equal[A]): Equal[Exp[A]] = ExpEqual(A)
-
-  // NB: Something like this currently needs to be defined for any Functor in
-  //     order to get the generalize operations for the algebra.
-  implicit def ToExpAlgebraOps[A](a: Algebra[Exp, A]): AlgebraOps[Exp, A] =
-    ToAlgebraOps[Exp, A](a)
-
-  implicit val ExpShow: Show ~> λ[α => Show[Exp[α]]] =
-    new (Show ~> λ[α => Show[Exp[α]]]) {
-      def apply[α](show: Show[α]) =
-        Show.show {
-          case Num(v)       => v.shows
-          case Mul(a, b)    =>
-            "Mul(" + show.shows(a) + ", " + show.shows(b) + ")"
-          case Var(s)       => "$" + s.name
-          case Lambda(p, a) => "Lambda(" + p.name + ", " + show.shows(a) + ")"
-          case Apply(f, a)  =>
-            "Apply(" + show.shows(f) + ", " + show.shows(a) + ")"
-          case Let(n, v, i) =>
-            "Let(" + n.name + ", " + show.shows(v) + ", " + show.shows(i) + ")"
-        }
-    }
-
-  implicit val ExpUnzip = new Unzip[Exp] {
-    def unzip[A, B](f: Exp[(A, B)]) = (f.map(_._1), f.map(_._2))
-  }
-}
-
-sealed trait Exp2[A]
-object Exp2 {
-  case class Const[A]() extends Exp2[A]
-  case class Num2[A](value: Int) extends Exp2[A]
-  case class Single[A](a: A) extends Exp2[A]
-
-  implicit val arbExp2: Arbitrary ~> λ[α => Arbitrary[Exp2[α]]] =
-    new (Arbitrary ~> λ[α => Arbitrary[Exp2[α]]]) {
-      def apply[α](arb: Arbitrary[α]): Arbitrary[Exp2[α]] =
-        Arbitrary(Gen.oneOf(
-          Gen.const(Const[α]),
-          Arbitrary.arbitrary[Int].map(Num2[α](_)),
-          arb.arbitrary.map(Single(_))))
-    }
-
-  def const = Fix[Exp2](Const())
-  def num2(v: Int) = Fix[Exp2](Num2(v))
-  def single(a: Fix[Exp2]) = Fix[Exp2](Single(a))
-
-  implicit val Exp2Functor: Functor[Exp2] = new Functor[Exp2] {
-    def map[A, B](fa: Exp2[A])(f: A => B): Exp2[B] = fa match {
-      case Const()   => Const[B]()
-      case Num2(v)   => Num2[B](v)
-      case Single(a) => Single(f(a))
-    }
-  }
-
-  implicit val Exp2Show: Show ~> λ[α => Show[Exp2[α]]] =
-    new (Show ~> λ[α => Show[Exp2[α]]]) {
-      def apply[α](show: Show[α]) =
-        Show.show {
-          case Const()   => "Const()"
-          case Num2(v)   => "Num2(" + v.shows + ")"
-          case Single(a) => "Single(" + show.shows(a) + ")"
-        }
-    }
-
-  implicit val Exp2Equal: Equal ~> λ[α => Equal[Exp2[α]]] =
-    new (Equal ~> λ[α => Equal[Exp2[α]]]) {
-      def apply[α](eq: Equal[α]) =
-        Equal.equal[Exp2[α]] {
-          case (Const(), Const())       => true
-          case (Num2(v1), Num2(v2))     => v1 ≟ v2
-          case (Single(a1), Single(a2)) => eq.equal(a1, a2)
-          case _                        => false
-        }
-    }
-  implicit def Exp2Equal2[A](implicit A: Equal[A]): Equal[Exp2[A]] = Exp2Equal(A)
-}
-
-
-class ExpSpec extends Spec {
-  import Exp._
-
-  implicit val arbExpInt: Arbitrary[Exp[Int]] = arbExp(Arbitrary.arbInt)
-  // NB: These are just a sanity check that the data structure created for the
-  //     tests is lawful.
-  checkAll(equal.laws[Exp[Int]])
-  checkAll(traverse.laws[Exp])
-}
-
-class Exp2Spec extends Spec {
-  import Exp2._
-
-  implicit val arbExp2Int: Arbitrary[Exp2[Int]] = arbExp2(Arbitrary.arbInt)
-  // NB: These are just a sanity check that the data structure created for the
-  //     tests is lawful.
-  checkAll(equal.laws[Exp2[Int]])
-  checkAll(functor.laws[Exp2])
-}
-
-class FixplateSpecs extends Specification with ScalaCheck with ScalazMatchers {
-  import Exp._
-
-  implicit def arbFix[F[_]]:
-      (Arbitrary ~> λ[α => Arbitrary[F[α]]]) => Arbitrary[Fix[F]] =
-    new ((Arbitrary ~> λ[α => Arbitrary[F[α]]]) => Arbitrary[Fix[F]]) {
-      def apply(FA: Arbitrary ~> λ[α => Arbitrary[F[α]]]):
-          Arbitrary[Fix[F]] =
-        Arbitrary(Gen.sized(size =>
-          FA(
-            if (size <= 0)
-              Arbitrary(Gen.fail[Fix[F]])
-            else
-              Arbitrary(Gen.resize(size - 1, arbFix(FA).arbitrary))).arbitrary.map(Fix(_))))
-    }
-
-  val example1ƒ: Exp[Option[Int]] => Option[Int] = {
-    case Num(v)           => v.some
-    case Mul(left, right) => (left ⊛ right)(_ * _)
-    case Var(v)           => None
-    case Lambda(_, b)     => b
-    case Apply(func, arg) => None
-    case Let(_, _, i)     => i
-  }
-
-  def addOneOptƒ[T[_[_]]]: Exp[T[Exp]] => Option[Exp[T[Exp]]] = {
-    case Num(n) => Num(n+1).some
-    case _      => None
-  }
-
-  def addOneƒ[T[_[_]]]: Exp[T[Exp]] => Exp[T[Exp]] =
-    orOriginal(addOneOptƒ)
-
+class MatryoshkaSpecs extends Specification with ScalaCheck with ScalazMatchers {
   def addOneOptExpExp2ƒ[T[_[_]]]: Exp[T[Exp2]] => Option[Exp2[T[Exp2]]] = {
-    case Num(n) => Exp2.Num2(n+1).some
+    case Num(n) => Num2(n+1).some
     case _      => None
   }
 
   def addOneExpExp2ƒ[T[_[_]]]: Exp[T[Exp2]] => Exp2[T[Exp2]] =
-    orDefault[Exp[T[Exp2]], Exp2[T[Exp2]]](Exp2.Const())(addOneOptExpExp2ƒ)
+    orDefault[Exp[T[Exp2]], Exp2[T[Exp2]]](exp2.Const())(addOneOptExpExp2ƒ)
 
   def addOneOptExp2Expƒ[T[_[_]]]: Exp2[T[Exp2]] => Option[Exp[T[Exp2]]] = {
-    case Exp2.Num2(n) => Num(n+1).some
-    case _            => None
+    case Num2(n) => Num(n+1).some
+    case _       => None
   }
 
   def addOneExp2Expƒ[T[_[_]]]: Exp2[T[Exp2]] => Exp[T[Exp2]] =
     orDefault[Exp2[T[Exp2]], Exp[T[Exp2]]](Num(0))(addOneOptExp2Expƒ)
 
-  def simplifyƒ[T[_[_]]: Recursive]: Exp[T[Exp]] => Option[Exp[T[Exp]]] = {
-    case Mul(a, b) => (a.project, b.project) match {
-      case (Num(0), Num(_)) => Num(0).some
-      case (Num(1), Num(n)) => Num(n).some
-      case (Num(_), Num(0)) => Num(0).some
-      case (Num(n), Num(1)) => Num(n).some
-      case (_,      _)      => None
-    }
-    case _         => None
-  }
-
-  def addOneOrSimplifyƒ[T[_[_]]: Recursive]: Exp[T[Exp]] => Exp[T[Exp]] = {
-    case t @ Num(_)    => addOneƒ(t)
-    case t @ Mul(_, _) => repeatedly(simplifyƒ[T]).apply(t)
-    case t             => t
-  }
-
   def extractLambdaƒ[T[_[_]]: Recursive]: Exp[(T[Exp], T[Exp2])] => Exp2[T[Exp2]] = {
     case Lambda(_, (exp, exp2)) => exp.project match {
-      case Num(a) => Exp2.Num2(a)
-      case _      => Exp2.Single(exp2)
+      case Num(a) => Num2(a)
+      case _      => Single(exp2)
     }
-    case _                      => Exp2.Const[T[Exp2]]
+    case _                      => exp2.Const[T[Exp2]]
   }
-
-  val MinusThree: Exp ~> Exp =
-    new (Exp ~> Exp) {
-      def apply[A](exp: Exp[A]): Exp[A] = exp match {
-        case Num(x) => Num(x-3)
-        case t      => t
-      }
-    }
 
   "Fix" should {
     "isLeaf" should {
@@ -394,17 +175,6 @@ class FixplateSpecs extends Specification with ScalaCheck with ScalazMatchers {
       }
     }
 
-    val eval: Algebra[Exp, Int] = {
-      case Num(x) => x
-      case Mul(x, y) => x*y
-      case _ => Predef.???
-    }
-
-    val findConstants: Exp[List[Int]] => List[Int] = {
-      case Num(x) => x :: Nil
-      case t      => t.fold
-    }
-
     "cata" should {
       "evaluate simple expr" in {
         val v = mul(num(1), mul(num(2), num(3)))
@@ -466,10 +236,6 @@ class FixplateSpecs extends Specification with ScalaCheck with ScalazMatchers {
       }
     }
 
-    def extractFactors: Coalgebra[Exp, Int] = x =>
-      if (x > 2 && x % 2 == 0) Mul(2, x/2)
-      else Num(x)
-
     "generalizeCoalgebra" should {
       "behave like ana" ! prop { (i: Int) =>
         i.apo(extractFactors.generalize[Fix[Exp] \/ ?]) must
@@ -478,13 +244,6 @@ class FixplateSpecs extends Specification with ScalaCheck with ScalazMatchers {
     }
 
     "topDownCata" should {
-      def subst[T[_[_]]: Recursive](vars: Map[Symbol, T[Exp]], t: T[Exp]):
-          (Map[Symbol, T[Exp]], T[Exp]) = t.project match {
-        case Let(sym, value, body) => (vars + ((sym, value)), body)
-        case Var(sym)              => (vars, vars.get(sym).getOrElse(t))
-        case _                     => (vars, t)
-      }
-
       "bind vars" in {
         val v = let('x, num(1), mul(num(0), vari('x)))
         v.topDownCata(Map.empty[Symbol, Fix[Exp]])(subst) must
@@ -496,14 +255,6 @@ class FixplateSpecs extends Specification with ScalaCheck with ScalazMatchers {
 
     "trans" should {
       // TODO
-    }
-
-    // Evaluate as usual, but trap 0*0 as a special case
-    def peval[T[_[_]]: Recursive](t: Exp[(T[Exp], Int)]): Int = t match {
-      case Mul((Proj(Num(0)), _), (Proj(Num(0)), _)) => -1
-      case Mul((_           , x), (_,            y)) => x * y
-      case Num(x)                                    => x
-      case _                                         => Predef.???
     }
 
     "attributePara" should {
@@ -613,10 +364,6 @@ class FixplateSpecs extends Specification with ScalaCheck with ScalazMatchers {
         }
       }
       "construct factorial" in {
-        def fact(x: Int): Exp[Fix[Exp] \/ Int] =
-          if (x > 1) Mul(-\/(num(x)), \/-(x-1))
-          else Num(x)
-
         4.apo(fact) must equal(mul(num(4), mul(num(3), mul(num(2), num(1)))))
       }
     }
@@ -624,8 +371,6 @@ class FixplateSpecs extends Specification with ScalaCheck with ScalazMatchers {
     "anamorphism" should {
       "pull out factors of two" in {
         "anaM" should {
-          def extractFactorsM(x: Int): Option[Exp[Int]] =
-            if (x == 5) None else extractFactors(x).some
           "pull out factors of two" in {
             12.anaM(extractFactorsM) must beSome(
               mul(num(2), mul(num(2), num(3)))
@@ -666,13 +411,6 @@ class FixplateSpecs extends Specification with ScalaCheck with ScalazMatchers {
       }
     }
 
-    def strings(t: Exp[(Int, String)]): String = t match {
-      case Num(x) => x.toString
-      case Mul((x, xs), (y, ys)) =>
-        xs + " (" + x + ")" + ", " + ys + " (" + y + ")"
-      case _ => Predef.???
-    }
-
     "zygo" should {
       "eval and strings" in {
         mul(mul(num(0), num(0)), mul(num(2), num(5))).zygo(eval, strings) must
@@ -708,8 +446,8 @@ class FixplateSpecs extends Specification with ScalaCheck with ScalazMatchers {
         case n => S(n - 1)
       })
 
-      case class Even(even: Boolean)
-      case class Odd(odd: Boolean)
+      final case class Even(even: Boolean)
+      final case class Odd(odd: Boolean)
 
       val isOdd: Nat[(Even, Odd)] => Odd = {
         case Z()             => Odd(false)
@@ -733,17 +471,6 @@ class FixplateSpecs extends Specification with ScalaCheck with ScalazMatchers {
       }
     }
 
-    // NB: This is better done with cata, but we fake it here
-    def partialEval[T[_[_]]: Corecursive: Recursive](t: Exp[Cofree[Exp, T[Exp]]]):
-        T[Exp] =
-      t match {
-        case Mul(x, y) => (x.head.project, y.head.project) match {
-          case (Num(a), Num(b)) => Num[T[Exp]](a * b).embed
-          case _                => t.map(_.head).embed
-        }
-        case _ => t.map(_.head).embed
-      }
-
     "histo" should {
       "eval simple literal multiplication" in {
         mul(num(5), num(10)).histo(partialEval[Fix]) must equal(num(50))
@@ -757,14 +484,6 @@ class FixplateSpecs extends Specification with ScalaCheck with ScalazMatchers {
           equal(lam('foo, mul(num(28), vari('foo))).convertTo[Mu])
       }
     }
-
-    def extract2and3(x: Int): Exp[Free[Exp, Int]] =
-      // factors all the way down
-      if (x > 2 && x % 2 == 0) Mul(Free.point(2), Free.point(x/2))
-      // factors once and then stops
-      else if (x > 3 && x % 3 == 0)
-        Mul(Free.liftF(Num(3)), Free.liftF(Num(x/3)))
-      else Num(x)
 
     "futu" should {
       "factor multiples of two" in {
@@ -866,5 +585,5 @@ class FixplateSpecs extends Specification with ScalaCheck with ScalazMatchers {
     }
   }
 
-  def expGen = Gen.resize(100, arbFix(arbExp).arbitrary)
+  def expGen = Gen.resize(100, corecursiveArbitrary(Exp.arbitrary).arbitrary)
 }
